checking for pairs

+++ poker_sim/src/models/card_model.rs
@@ -1,5 +1,5 @@
 #[derive(Copy, Debug, Clone, PartialEq)]
 pub struct Card {
-    pub suit: u32, 
-    pub rank: u32
+    pub suit: usize, 
+    pub rank: usize
 }
